requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
transformers>=4.41.0
huggingface-hub>=0.23.0
pydantic>=2.7.0
pillow>=10.3.0
safetensors>=0.4.3
